divided whole project into modules
