+++ backend/BusinessLogic/src/main/java/com/example/demo/Duel/RowStatus.java
