+++ backend/BusinessLogic/src/main/java/com/example/demo/Utils.java
@@ -1,0 +1,11 @@
+package com.example.demo;
+
+import java.util.Random;
+
+public class Utils {
+    public static int getRandomNumber(int min, int max) {
+        Random random = new Random();
+        return random.nextInt(max - min +1) + min;
+    }
+
+}
