--- backend/src/main/java/com/example/demo/Utils.java
@@ -1,14 +1,0 @@
-package com.example.demo;
-
-import com.example.demo.CardsServices.CardDisplay;
-import com.example.demo.CardsServices.Cards.Card;
-
-import java.util.Random;
-
-public class Utils {
-    public static int getRandomNumber(int min, int max) {
-        Random random = new Random();
-        return random.nextInt(max - min +1) + min;
-    }
-
-}
