--- backend/src/test/java/com/example/demo/DuelTests/TestDuelAfterDeal.java
@@ -1,281 +1,0 @@
-package com.example.demo.DuelTests;
-
-import com.example.demo.CardsServices.CardDisplay;
-import com.example.demo.CardsServices.CardsParser;
-import com.example.demo.Consts;
-import com.example.demo.Duel.CardDuel;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import static com.example.demo.CardsServices.Cards.CardsFactory.*;
-import static org.junit.jupiter.api.Assertions.*;
-import static com.example.demo.TestsData.TestConsts.*;
-import static com.example.demo.TestsData.TestsUtils.*;
-
-class TestDuelAfterDeal {
-    CardDuel duel;
-
-
-    @BeforeEach
-    public void setUp(){
-        duel = CardDuel.createDuel();
-        duel.registerPlayerToDuel(secondPlayer);
-        duel.registerPlayerToDuel(firstPlayer);
-        List<CardDisplay> allCards = CardsParser.getCardsDisplay(createAllCards());
-        List<CardDisplay> noEffectCards = allCards.stream()
-                .filter(c -> c.getCardInfo().equals(""))
-                .collect(Collectors.toList());
-        duel.parseCardsFor( noEffectCards , firstPlayer);
-        duel.parseCardsFor( noEffectCards , secondPlayer);
-        duel.dealCards();
-    }
-
-    @Test
-    public void beforePlayingCard_boardIsEmpty(){
-        assertTrue(duel.getRowOf(firstPlayer, Consts.firstRow).isEmpty());
-    }
-
-    @Test
-    public void afterPlayingCard_boardIsNotEmpty(){
-        int cardsInHandBeforePlaying = duel.getHandOf(firstPlayer).size();
-        playCardWithoutTargeting(duel, duel.getHandOf(firstPlayer).get(0), Consts.firstRow, firstPlayer);
-        assertFalse(duel.getRowOf(firstPlayer, Consts.firstRow).isEmpty());
-        assertNotEquals(cardsInHandBeforePlaying, duel.getHandOf(firstPlayer).size());
-    }
-
-    @Test
-    public void clearBoard_hasNoPoints(){
-        assertEquals(0, getBoardPointsOf(firstPlayer, duel));
-    }
-
-    @Test
-    public void onFirstPlayerTurn_player2CantPlayCard(){
-        playCardWithoutTargeting(duel, duel.getHandOf(secondPlayer).get(0), Consts.firstRow, secondPlayer);
-        assertTrue(duel.getRowOf(secondPlayer, Consts.firstRow).isEmpty());
-    }
-
-    @Test
-    public void afterPlayingCardsWithPoints_theSamePointsAreOnBoard(){
-        CardDisplay cardToPlay_first = duel.getHandOf(firstPlayer).get(0);
-        CardDisplay cardToPlay_second = duel.getHandOf(firstPlayer).get(1);
-        playCardWithoutTargeting(duel, cardToPlay_first, Consts.firstRow, firstPlayer);
-        playCardWithoutTargeting(duel, duel.getHandOf(secondPlayer).get(0), Consts.firstRow, secondPlayer);
-        playCardWithoutTargeting(duel, cardToPlay_second, Consts.secondRow, firstPlayer);
-        assertEquals(cardToPlay_first.getPoints() + cardToPlay_second.getPoints(), getBoardPointsOf(firstPlayer, duel));
-    }
-
-    @Test
-    public void afterPlayingCard_cantPlayAnother(){
-        int expectedCardsOnBoard = 1;
-        playCardWithoutTargeting(duel, duel.getHandOf(firstPlayer).get(0), Consts.firstRow, firstPlayer);
-        assertEquals(expectedCardsOnBoard, duel.getRowOf(firstPlayer, Consts.firstRow).size());
-
-        playCardWithoutTargeting(duel, duel.getHandOf(firstPlayer).get(0), Consts.firstRow, firstPlayer);
-        assertEquals(expectedCardsOnBoard, duel.getRowOf(firstPlayer, Consts.firstRow).size());
-
-        playCardWithoutTargeting(duel, duel.getHandOf(secondPlayer).get(0), Consts.firstRow, secondPlayer);
-        playCardWithoutTargeting(duel, duel.getHandOf(secondPlayer).get(0), Consts.firstRow, secondPlayer);
-        assertEquals(expectedCardsOnBoard, duel.getRowOf(secondPlayer, Consts.firstRow).size());
-    }
-
-    @Test
-    public void playerCanEndRound(){
-        duel.endRoundFor(firstPlayer);
-        assertTrue(duel.didEndRound(firstPlayer));
-    }
-
-    @Test
-    public void atStartOfGame_nobodyEndedRound(){
-        assertFalse(duel.didEndRound(firstPlayer));
-        assertFalse(duel.didEndRound(secondPlayer));
-    }
-
-    @Test
-    public void afterOnePlayerEndsRound_turnsDontSwitch() {
-        playCardWithoutTargeting(duel, duel.getHandOf(firstPlayer).get(0), Consts.firstRow, firstPlayer);
-        duel.endRoundFor(secondPlayer);
-        playCardWithoutTargeting(duel, duel.getHandOf(firstPlayer).get(0), Consts.firstRow, firstPlayer);
-        playCardWithoutTargeting(duel, duel.getHandOf(firstPlayer).get(0), Consts.firstRow, firstPlayer);
-        int expectedNumberOfCardsOnBoard = 3;
-        assertEquals(expectedNumberOfCardsOnBoard, duel.getRowOf(firstPlayer, Consts.firstRow).size());
-    }
-
-    @Test
-    public void afterTwoPlayerEndRound_newRoundStarts(){
-        startNextRound_firstPlayerFirst();
-        assertFalse(duel.didEndRound(firstPlayer));
-        assertFalse(duel.didEndRound(secondPlayer));
-    }
-
-    @Test
-    public void afterEndingFirstRound_newCardsAreDealt() {
-        int cardsInHandBeforeNewTurn = duel.getHandOf(firstPlayer).size();
-        startNextRound_firstPlayerFirst();
-        int cardsInHandAfterNewTurn = duel.getHandOf(firstPlayer).size();
-        assertNotEquals(cardsInHandAfterNewTurn, cardsInHandBeforeNewTurn);
-    }
-
-    @Test
-    public void atTheGameStart_noOneWonRound(){
-        assertEquals(0, duel.getWonRoundsOf(firstPlayer));
-        assertEquals(0, duel.getWonRoundsOf(secondPlayer));
-    }
-
-    @Test
-    public void afterWinningTurn_onePlayerScoresPoints(){
-        playCardAsFirstPlayer_startNextRound();
-        assertEquals(1, duel.getWonRoundsOf(firstPlayer));
-        assertEquals(0, duel.getWonRoundsOf(secondPlayer));
-    }
-
-    @Test
-    public void afterNewRound_boardIsEmpty(){
-        playCardWithoutTargeting(duel, duel.getHandOf(firstPlayer).get(0), Consts.secondRow, firstPlayer);
-        duel.endRoundFor(secondPlayer);
-        playCardWithoutTargeting(duel, duel.getHandOf(firstPlayer).get(0), Consts.thirdRow, firstPlayer);
-        playCardWithoutTargeting(duel, duel.getHandOf(firstPlayer).get(0), Consts.firstRow, firstPlayer);
-        duel.endRoundFor(firstPlayer);
-        for(int row = 0 ; row < Consts.rowsNumber ; ++row){
-            assertTrue(duel.getRowOf(firstPlayer, row).isEmpty());
-        }
-    }
-
-    @Test
-    public void afterTwoRoundsWon_playerWonGame(){
-        playCardAsFirstPlayer_startNextRound();
-        playCardAsFirstPlayer_startNextRound();
-        assertTrue(duel.didWon(firstPlayer));
-    }
-
-    @Test
-    public void afterWonGame_cantPlayCard(){
-        playCardAsFirstPlayer_startNextRound();
-        playCardAsFirstPlayer_startNextRound();
-        int cardsOnBoard_beforePlay = duel.getRowOf(firstPlayer, Consts.firstRow).size();
-        playCardWithoutTargeting(duel, duel.getHandOf(firstPlayer).get(0), Consts.firstRow, firstPlayer);
-        int cardsOnBoard_afterPlay = duel.getRowOf(firstPlayer, Consts.firstRow).size();
-        assertEquals(cardsOnBoard_afterPlay, cardsOnBoard_beforePlay);
-    }
-
-    public void playCardAsFirstPlayer_startNextRound(){
-        playCardWithoutTargeting(duel, duel.getHandOf(firstPlayer).get(0), Consts.firstRow, firstPlayer);
-        startNextRound_secondPlayerFirst();
-    }
-
-    public void startNextRound_secondPlayerFirst(){
-        duel.endRoundFor(secondPlayer);
-        duel.endRoundFor(firstPlayer);
-    }
-    public void startNextRound_firstPlayerFirst(){
-        duel.endRoundFor(firstPlayer);
-        duel.endRoundFor(secondPlayer);
-    }
-
-    @Test
-    public void testCalculatingWonRounds(){
-        CardDisplay morePointsCard = Collections.max(
-                duel.getHandOf(firstPlayer), Comparator.comparingInt(CardDisplay::getPoints)
-        );
-        CardDisplay lessPointsCard = Collections.min(
-                duel.getHandOf(secondPlayer), Comparator.comparingInt(CardDisplay::getPoints)
-        );
-        playCardWithoutTargeting(duel, morePointsCard, Consts.firstRow, firstPlayer);
-        playCardWithoutTargeting(duel, lessPointsCard, Consts.firstRow, secondPlayer);
-        duel.endRoundFor(firstPlayer);
-        duel.endRoundFor(secondPlayer);
-
-        assertEquals(1, duel.getWonRoundsOf(firstPlayer));
-        assertEquals(0, duel.getWonRoundsOf(secondPlayer));
-    }
-
-    @Test
-    public void afterTie_bothPlayersScoresPoint(){
-        CardDisplay theSamePoints_firstPlayer = duel.getHandOf(firstPlayer).get(0);
-        CardDisplay theSamePoints_secondPlayer = duel.getHandOf(secondPlayer).stream()
-                .filter(c -> c.getPoints() == theSamePoints_firstPlayer.getPoints())
-                .findFirst().orElse(null);
-        playCardWithoutTargeting(duel, theSamePoints_firstPlayer, Consts.firstRow, firstPlayer);
-        playCardWithoutTargeting(duel, theSamePoints_secondPlayer, Consts.firstRow, secondPlayer);
-        duel.endRoundFor(firstPlayer);
-        duel.endRoundFor(secondPlayer);
-
-        assertEquals(1, duel.getWonRoundsOf(firstPlayer));
-        assertEquals(1, duel.getWonRoundsOf(secondPlayer));
-    }
-
-    @Test
-    public void afterCardDies_goesToGraveyard() {
-        duel = createDuel(List.of(minion, archer));
-        List<CardDisplay> hand1 = duel.getHandOf(firstPlayer);
-        List<CardDisplay> hand2= duel.getHandOf(secondPlayer);
-        playCardWithoutTargeting(duel, findByName(hand1, minion), firstRow, firstPlayer);
-        playCardWithCardTargeting(duel, findByName(hand2, archer), firstRow, findByName(hand1,minion ), secondPlayer);
-        assertEquals(1, duel.getGraveyardOf(firstPlayer).size());
-
-    }
-    @Test
-    public void afterCardIsBurned_goesToGraveyard() {
-        duel = createDuel(List.of(knight, conflagration));
-        List<CardDisplay> hand1 = duel.getHandOf(firstPlayer);
-        List<CardDisplay> hand2= duel.getHandOf(secondPlayer);
-        playCardWithoutTargeting(duel, findByName(hand1, knight), firstRow, firstPlayer);
-        playSpecialCardWithoutTargeting(duel, findByName(hand2, conflagration), secondPlayer);
-        assertEquals(1, duel.getGraveyardOf(firstPlayer).size());
-    }
-    @Test
-    public void afterNewRound_everyCardGoesToGraveyard() {
-        duel = createDuel(List.of(viking, knight));
-
-        List<CardDisplay> hand1 = duel.getHandOf(firstPlayer);
-        List<CardDisplay> hand2= duel.getHandOf(secondPlayer);
-        playCardWithoutTargeting(duel, findByName(hand1, viking), firstRow, firstPlayer);
-        playCardWithoutTargeting(duel, findByName(hand2, viking), firstRow, secondPlayer);
-        playCardWithoutTargeting(duel, findByName(hand1, knight), firstRow, firstPlayer);
-        playCardWithoutTargeting(duel, findByName(hand2, knight), firstRow, secondPlayer);
-        duel.endRoundFor(firstPlayer);
-        duel.endRoundFor(secondPlayer);
-
-        assertTrue(duel.getRowOf(firstPlayer, firstRow).isEmpty());
-        assertTrue(duel.getRowOf(secondPlayer, firstRow).isEmpty());
-
-        assertEquals(2,duel.getGraveyardOf(firstPlayer).size());
-        assertEquals(2,duel.getGraveyardOf(secondPlayer).size());
-    }
-
-    @Test
-    public void testPlayingTwoSameCards() {
-        duel = createDuel(List.of(viking, viking));
-        List<CardDisplay> handCards_firstPlayer = duel.getHandOf(firstPlayer);
-        List<CardDisplay> handCards_secondPlayer = duel.getHandOf(secondPlayer);
-        playCardWithoutTargeting(duel,handCards_firstPlayer.get(0), firstRow, firstPlayer);
-        playCardWithoutTargeting(duel,handCards_secondPlayer.get(0), firstRow, secondPlayer);
-        playCardWithoutTargeting(duel,handCards_firstPlayer.get(1), firstRow, firstPlayer);
-        playCardWithoutTargeting(duel,handCards_secondPlayer.get(1), firstRow, secondPlayer);
-        List<CardDisplay> rowCards = duel.getRowOf(firstPlayer, firstRow);
-        assertNotEquals(rowCards.get(0).getId(), rowCards.get(1).getId());
-    }
-
-    @Test
-    public void testDealingDamageToOneOfTwoSameCards() {
-        duel = createDuel(List.of(viking, viking, archer));
-        List<CardDisplay> handCards_firstPlayer = duel.getHandOf(firstPlayer);
-        List<CardDisplay> handCards_secondPlayer = duel.getHandOf(secondPlayer);
-        playCardWithoutTargeting(duel, handCards_firstPlayer.get(0), firstRow, firstPlayer );
-        playCardWithoutTargeting(duel, handCards_secondPlayer.get(0), firstRow, secondPlayer);
-        playCardWithoutTargeting(duel, handCards_firstPlayer.get(1), firstRow, firstPlayer );
-
-        List<CardDisplay> enemyBoard = duel.getRowOf(firstPlayer, firstRow);
-        playCardWithCardTargeting(duel, handCards_secondPlayer.get(2), firstRow,enemyBoard.get(1), secondPlayer);
-        enemyBoard = duel.getRowOf(firstPlayer, firstRow);
-        assertEquals(viking.getPoints() - archerDamage, enemyBoard.get(1).getPoints());
-        assertEquals(viking.getPoints(), enemyBoard.get(0).getPoints());
-    }
-
-
-
-}
